+++ CHANGELOG.md
@@ -15,7 +15,7 @@
   **Breaking**: update all zoom level dependent styles by adding 1 to each zoom level.
 - Added `Point` class to encapsulate x, y coords.
   API methods will accept either `[x, y]` or `new llmr.Point(x, y)`.
-- Changed `Map` `panBy` signature to accept `point` instead of `x, y`.
+- Changed `Map` `panBy` and `Map`/`Source` `featuresAt` to accept `point` instead of `x, y`.
 - Changed `Map` `fitBounds` and `zoomPanTo` to accept `offset` (Point) instead of `offsetX, offsetY`.
 - Raster buckets now need to specify `type: 'raster'` in the stylesheet.